+++ packages/sdk/src/models/mutations.ts
@@ -324,7 +324,7 @@ class Mutations {
             }
 
             case MutationTypes.UPDATE_SINGLE_FIELD_CONFIG: {
-                const {tableId, id, config} = mutation;
+                const {tableId, id, config, opts} = mutation;
                 const table = this._base.getTableByIdIfExists(tableId);
                 if (!table) {
                     throw spawnError("Can't update field: No table with id %s exists", tableId);
@@ -346,6 +346,7 @@ class Mutations {
                     currentConfig,
                     field._data,
                     billingPlanGrouping,
+                    opts,
                 );
 
                 if (!validationResult.isValid) {