+++ packages/sdk/test/models/field.test.ts
@@ -69,6 +69,7 @@ describe('Field', () => {
                             choices: [{name: 'pick me'}],
                         },
                     },
+                    opts: {},
                 },
                 {holdForMs: 100},
             );